Add files via upload
#Daily Question 
<Foundation>
Assume you are a volunteer at the sports meet of The Hong Kong University of Science and Technology Guangzhou (HKUSTGZ). There are n athletes, each with a unique number from 1 to n, but the order in which they check in is disorganized. You need to design a program to sort these numbers in ascending order on the check-in machine, and thereafter, be able to use the binary search method to find a specific number.